MyPy + Jupyter Notebooks with Testing + Handle Txn Params (#17)
# `mypy`
Enabling mypy by incorporating #16 and making other similar changes to new code. This makes it possible for downstream dependencies that incorporate graviton to be mypy-compliant as well.

# Allow handling of Transaction Params in Logic Sig and App Dry-Runs
Previously, graviton was using the old Python SDK DryRunHelper which hard-coded the transaction information associated with a dry run. This made it impossible to write logic sigs that vallidated the transaction information in their program. This PR fixes this situation. In fact, the image below was only possible to create thanks to this new addition.

# Jupyter Notebooks
Introducing notebooks/quadratic_factoring_game.ipynb which shows how to leverage graviton for creating graphs in jupyter such as below. The notebook also runs during integration testing.